Enable authn and authz in Fedora (#58)
* Enable authn and authz in Fedora

- Use fcrepo-webapp-plus instead of stock fedora, as it has webac built in
- Configure Fedora (tomcat) for authn
- Configure apix, camel, and extensions to use authn
- Update exercises in guide to add username/password when required
- Use shell script to bootstrap Amherst examples, rather than self-loading

* Properly handle authn in rdfvis php extension

Easyrdf was making unauthenticated requests to Fedora.  This fixes the
situation so that the _provided_ credentials (all headers, really) are
used to formulate the request for a given resource.

* Refactor config for public URIs

Use new env variables PUBLIC_REPOSITORY_* for specifying the public
repository endpoint.  docker-machine (or DNS) users will modify these
variables to suit their local machine IP or DNS name.

Update various images to use these environment variables

* Forward all request and response headers in rdfvis

This allows the proper authentication-related headers to pass from
the client to Fedora (and vice versa) in order to allow authenticated
access to the RDF serializations of resources

* Use PUBLIC_REPOSITORY_HOST for default auth host
* Use PUBLIC_REPOSITORY_HOST for indexing auth host

Indexing would fail on docker-machine because the auth host always
defaulted to localhost